Move all common graph helpers to common place
